16.2.Upgrading/Migrating from StofDoctrineExtensions{Changing to antishov/doctrine-extensions-bundle: composer require antishov/doctrine-extensions-bundle}
Now go back to the homepage of the fork, find the composer require line, copy it, and re-install the library:

composer require antishov/doctrine-extensions-bundle
This basically gives us the same library but at a newer version. The author also created an identical recipe for this package, so even the recipe gets re-installed nicely.

Commit the files we know we want to keep:

git add composer.json composer.lock symfony.lock

Now selectively-choose the changes from the updated recipe by running:

git add -p
For bundles.php - it looks like it removed the bundle... but if you hit "y", it just moved it. A meaningless change. And next, because it re-installed the recipe, it removed our custom changes. Hit "n" to skip those.

Let's commit!

git commit -m "using doctrine extensions bundle fork"
And then, revert the changes to the config file:

git checkout .
So... that update was weird. Let's close some tabs and refresh. Yas! The deprecations jumped from 25 to 16.

We're killing it! The next deprecations are going to uncover that we also need to upgrade DoctrineBundle... from version 1 to 2 - a significant jump.