17.2.Upgrading to DoctrineBundle 2.0
So probably we need to go into composer.json and update its version constraint. Search for doctrine/doctrine-bundle. Huh... it's actually not inside our composer.json file. It must be a dependency of something else. Let's find out more. Run:

composer why doctrine/doctrine-bundle
Ok: doctrine-bundle is required by both fixtures-bundle and migrations-bundle. But the original reason that it was installed was because of symfony/orm-pack, which allows version 1 or 2. Remember: orm-pack is a sort of "fake" library that requires other libraries. It gives us an easy way to install Doctrine and some other Doctrine-related libraries.
I want to have more control over doctrine/doctrine-bundle so that I can force version 2 to be used, instead of just 1 or 2. To do that, we can "unpack" the pack. Run:

composer unpack symfony/orm-pack

Very simply, this removes symfony/orm-pack from our composer.json file. And replaces it with the libraries that it requires.