17.4.Upgrading to DoctrineBundle 2.0{Handling Composer Update Problems}
Handling Composer Update Problems
Now that we're forcing version 2, let's... see if it will update! Run:

composer update doctrine/doctrine-bundle
And... this does not work. I should've seen this coming. doctrine-bundle can't be updated to version 2 because our project currently has doctrine-fixtures-bundle 3.2, which requires doctrine-bundle 1.6. So apparently we also need to update doctrine-fixtures-bundle. Ok! Copy the library name and say:

composer update doctrine/doctrine-bundle doctrine/doctrine-fixtures-bundle
We may need to change that bundle's version constraints to allow it to upgrade to a new major version.
I honestly don't really know. My hope is that a newer 3.something version will allow doctrine-bundle 2.0. But when we check on Composer... bah! It didn't work! But this time because of a different library: doctrine-migrations-bundle. That also needs to be updated. No problem: copy its name and add it to the end of our composer update line.
We're now allowing doctrine-bundle, doctrine-fixtures-bundle and doctrine-migrations-bundle all to update and... it still doesn't work. Sheesh. Let's see: this time it's because doctrine-migrations-bundle requires doctrine/migrations 2.2 and apparently we're locked at a lower version of that. Yea we have 2.1.1.
It's the same problem... again. Well, it's slightly different. We could add doctrine/migrations to the end of our composer update command - or even use doctrine/* - and try it again. That should work.
Or we can add --with-dependencies. This says: allow any of these three bundles to update and allow their dependencies to update. doctrine/migrations is not in our composer.json file: it's a dependency of doctrine-migrations-bundle. Oh, and if you really want the easy way out, we could have just ran composer update with no arguments to allow anything to update. But I prefer to update as little as possible at one time. Try the command
composer update doctrine/doctrine-bundle                 doctrine/doctrine-fixtures-bundle                 doctrine/doctrine-migrations-bundle                 --with-dependencies
And... it worked! Yeeeea! Then... exploded at the bottom. We'll talk about that in a minute. First, look back up: it upgraded doctrine-migrations-bundle and doctrine-fixtures-bundle both to new minor versions. So there shouldn't be any breaking changes in those.
The doctrine-bundle upgrade was over a major version - from 1 to 2 - so it shouldn't be a huge surprise that it made our code go bonkers: the new version does have some breaking changes.

Exit: DoctrineCacheBundle
One other thing I want to point out is that doctrine-cache-bundle was removed. That's no longer needed by doctrine and you shouldn't use it anymore either: use Symfony's cache.
Next, let's fix our app to work with DoctrineBundle 2.0 and update its recipe, which contains a few important config changes.