19.1.Fixing our Deprecations: Form, Controller & Mailer{Form getExtendedTypes() Deprecation}
We are now super close to fixing all the deprecation warnings that block us from going to Symfony 5. Let's check out the current list for the homepage. There are technically 12 deprecations. But remember, we can ignore all the ones from doctrine/persistence because they're not related to Symfony.

Form getExtendedTypes() Deprecation
With that in mind... if you look closely, there are really only two real deprecations left... and they look like the same thing: something about TextareaSizeExtension should implement a static getExtendedTypes() method.
This TextareaSizeExtension class is a "form type extension" that we built in an earlier tutorial. Let's go check it out: src/Form/TypeExtension/TextareaSizeExtension.php
And... PhpStorm is immediately mad at us:

  Class must be declared abstract or implement method getExtendedTypes().

This is the error you see when you have a class that implements an interface but is missing one of the methods that the interface requires. But in this case, that's not technically true. Hold command or control and click the interface to jump to that file.
In reality, there is no getExtendedTypes() method on this interface! It has getExtendedType() - that's the old, deprecated method - but no getExtendedTypes(). It's not actually on the interface, it's just described on top of the class in comments.
You're seeing Symfony's deprecation system in action. If Symfony suddenly added this new getExtendedTypes() method to the interface in Symfony 4.4, it would have broken our app when we upgraded. That would violate Symfony's backwards-compatibility promise... which basically says: we will not break your app on a minor upgrade.
Instead Symfony describes that you need this method and warns you to add it via the deprecation system. It will be added to the interface in Symfony 5.0. Our job is to add this new static getExtendedTypes() method that returns iterable.
We got this! At the bottom of our class, add public static function getExtendedTypes() with an iterable return type. Inside, return an array with the same class as the old method.
As soon as we do this, the old, getExtendedType() method won't be called anymore.
And it will be gone from the interface in Symfony 5.0. But we do need to keep it temporarily... because, again, for backwards compatibility, it does still exist on the interface in Symfony 4.4. If we removed it from our class, PHP would be super angry. I'll add a comment:

  not used anymore - remove in 5.0

Cool! Let's go close the profiler, refresh and open the new deprecations list. And... hey! Ignoring the doctrine/persistence stuff, our homepage is now free of deprecations!

Does that mean our app is ready for Symfony 5? Ah... not so fast: we still need to do a few more things to be sure that no deprecated code is hiding.