19.3.Fixing our Deprecations: Form, Controller & Mailer{Mailer: NamedAddress to Address}
Mailer: NamedAddress to Address
Ok! Another deprecation down and the homepage is once again not triggering any deprecated code. Let's surf around and see if we notice any other deprecations... how about the registration page: SymfonyNerd@example.com, any password, agree to the terms and... woh! That's not a deprecation... that's a huge error!
In theory, you should never get an error after a "minor" version upgrade - like Symfony 4.3 to 4.4. But this is coming from Symfony's Mime component, which is part of Mailer. And because Mailer was experimental until Symfony 4.4 there were some breaking changes from 4.3 to 4.4. We saw this one mentioned earlier when we looked at the Mailer CHANGELOG. Basically, NamedAddress is now called Address.

Where do we use NamedAddress? Let's find out! At your terminal, my favorite way to find out is to run:

  git grep NamedAddress

It's used in SetFromListener, Mailer and MailerTest. Let's do some updating. I'll start with src/Service/Mailer.php: change the use statement from NamedAddress to Address, then search for NamedAddress and remove the Named part here and in one other place. Next is EventListener/SetFromListener.
Make the same change on top... and below.
The last place is inside of tests/: Service/MailerTest. Let's see: remove Named from the use statement... and then it's used below in 2 places.
Got it! Let's try the registration page now: refresh and... validation error. Change to a new email, agree to the terms and... got it!
Ok, the deprecations are gone from the homepage and registration page at least. Are we done? How can we be sure?
Next, let's use a few more tricks to really be sure the deprecations are gone.