21.1.Upgrading to Symfony 5.0{Changing composer.json for Symfony 5.0; Updating Symfony Packages in Composer}
We've done it! We fixed all the deprecations in our app... except for the doctrine/persistence stuff, which we don't need to worry about because we're not upgrading that library. That means... we are ready for Symfony5!

Changing composer.json for Symfony 5.0
How... do we actually upgrade? We already know: it's the exact same process we used to upgrade from 4.3 to 4.4.
Open up composer.json. Our goal is to update all of these symfony/ libraries to 5.0. Well, not quite all of them - a few are not part of the main Symfony library, like monolog-bundle. But basically, everything that has 4.4.* now needs to be 5.0.*.
We also need to update one more thing: the extra.symfony.require value. This is primarily a performance optimization that helps Composer filter out extra Symfony versions when it's trying to resolve packages. This also needs to change to 5.0.*.
Let's... do it all at once: Find 4.4.*, replace it with 5.0.* and hit "Replace all". And then... make sure that this didn't accidentally replace any non-Symfony packages that may have had the same version.... looks good.

Updating Symfony Packages in Composer
We're ready! At your terminal... I'll hit Ctrl+C to stop the log tail.... and run the same command we used when upgrading from Symfony 4.3 to 4.4:

composer update symfony/*
That's it! It's that easy! We're done! Kidding - it's never that easy: you will almost definitely get some dependency errors. Probably... several. Ah, here's our first.