22.2.Secrets Management Setup{Converting an Env Var to a Secret}
Converting an Env Var to a Secret
Ok, as soon as you have an environment variable that you want to convert to a secret, you need to fully remove it as an environment variable: do not set it as an environment variable anywhere anymore. I'll remove MAILER_DSN from .env and if we were overriding it in .env.local, I would also remove it from there.
Not surprisingly, when you refresh, we're greeted with a great big ugly error:

  The environment variable is not found.