23.2.Production Secrets{Seeing the prod Secret Value}
Seeing the prod Secret Value
Let's go make sure this whole prod vault idea works. Right now, if we refresh the page, it still shows us the null value because we are still in the dev environment.
Open up your .env file and, temporarily, change APP_ENV to prod.
Then, find your console and clear the cache:

php bin/console cache:clear
I don't need to add --env=prod now because we are already in the 'prod environment thanks to the APP_ENV' change. Ok, go try it! Refresh and... yes! That's the value from the prod vault! Symfony automatically used the private key to decrypt it.
And if the Decrypt Key is Missing?
What would happen if the decrypt key wasn't there? Let's find out! Temporarily delete the decrypt key - but make sure you can get it back: if you lose this key, you won't ever be able to decrypt your secrets and you'll need to create a new private key and re-add them all again. That would be... a bummer.

Refresh now to see... oh! Giant 500 page... but we can't see the error. Check out the logs:

tail var/log/prod.log
And... there it is:

  Environment variable not found: "MAILER_DSN".